services/intraday/pyproject.toml
[project]
name = "ecs-intraday-service"
version = "0.1.0"
description = "Energy Control System - Intraday Service"
requires-python = ">=3.11"

dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "psycopg[binary]",
    "pydantic",
    "kafka-python",
    "python-dotenv",
    "ortools"
]
